fix(islamic-coin): use dynamic loading to fix hydration errors
